move exception traits update changelog
