+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [v.0.5.3](https://github.com/upb-uc4/hyperledger_api/compare/v0.5.2...v0.5.3) (2020-08-11)
+
+## Feature
+- Update Exceptions according to API definition
+
+## Bug Fixes
+- 
+
+## Refactor
+- 
+
+### Usability
+- 
+
 # [v.0.5.2](https://github.com/upb-uc4/hyperledger_api/compare/v0.5.1...v0.5.2) (2020-08-07)
 
 ## Feature
