+++ src/main/scala/de/upb/cs/uc4/hyperledger/exceptions/TransactionException.scala
@@ -1,5 +1,7 @@
 package de.upb.cs.uc4.hyperledger.exceptions
 
+import de.upb.cs.uc4.hyperledger.exceptions.traits.TransactionExceptionTrait
+
 case class TransactionException(transactionId: String, jsonError: String) extends TransactionExceptionTrait{
   override def toString: String =
     s"The provided transaction: '${transactionId}' failed with an error: ${jsonError}"
