+++ src/main/scala/de/upb/cs/uc4/hyperledger/exceptions/UnhandledException.scala
@@ -1,5 +1,7 @@
 package de.upb.cs.uc4.hyperledger.exceptions
 
+import de.upb.cs.uc4.hyperledger.exceptions.traits.HyperledgerExceptionTrait
+
 case class UnhandledException(transactionId: String, innerException: Exception) extends HyperledgerExceptionTrait {
   override def toString(): String =
     "The provided transaction: \"" + transactionId + "\" failed with an unhandled exception:\n" + innerException
