+++ src/main/scala/de/upb/cs/uc4/hyperledger/exceptions/traits/HyperledgerExceptionTrait.scala
@@ -1,6 +1,7 @@
-package de.upb.cs.uc4.hyperledger.exceptions
+package de.upb.cs.uc4.hyperledger.exceptions.traits
 
 trait HyperledgerExceptionTrait extends Exception {
   def transactionId: String
+
   def innerException: Exception
 }
